:art: SplitImportFunc
